+++ contracts/scripts/deploy.js
@@ -1,4 +1,8 @@
+const gasLimit = "20000000";
 process.env.CONFIRMATIONS = 1; // required for deployProtocolConfigFacet()
+process.env.DEPLOYER_GAS_LIMIT_TEST = gasLimit;
+process.env.AUTH_TOKEN_OWNERS_LOCAL = "";
+process.env.TIP_MULTIPLIER = "10";
 /* eslint @typescript-eslint/no-var-requires: "off" */
 const hre = require("hardhat");
 const ethers = hre.ethers;
@@ -23,12 +27,11 @@ const {
   deploymentComplete
 } = require("../protocol-contracts/scripts/util/utils");
 const {
-  deployMockTokens
+  deployMockTokens,
+  deployAndMintMockNFTAuthTokens
 } = require("../protocol-contracts/scripts/util/deploy-mock-tokens");
 const { oneMonth } = require("../protocol-contracts/test/util/constants");
 const AuthTokenType = require("../protocol-contracts/scripts/domain/AuthTokenType");
-
-const gasLimit = "20000000";
 
 function getConfig(network) {
   return [
@@ -51,6 +54,12 @@ function getConfig(network) {
  * @returns {lensAddress: string, ensAddress: string}
  */
 function getAuthTokenContracts(network) {
+  if (network === "localhost") {
+    return {
+      lensAddress: "0x2bdCC0de6bE1f7D2ee689a0342D76F52E8EFABa3",
+      ensAddress: "0x7969c5eD335650692Bc04293B07F5BF2e7A673C0"
+    };
+  }
   return {
     lensAddress: authTokenAddresses.LENS[network],
     ensAddress: authTokenAddresses.ENS[network]
@@ -81,6 +90,7 @@ function getNoArgFacetNames() {
 }
 
 async function main() {
+  console.log("compile");
   // Compile everything (in case run by node)
   await hre.run("compile");
 
@@ -113,7 +123,7 @@ async function main() {
   if (hre.network.name === "localhost" || !config[0].token) {
     console.log(`\n💎 Deploying mock tokens...`);
     const [mockBosonToken] = await deployMockTokens(gasLimit);
-    deploymentComplete("BosonToken", mockBosonToken.address, [], contracts);
+    deploymentComplete("BosonToken", mockBosonToken.address, [], "", contracts);
     config[0].token = mockBosonToken.address;
   }
 
@@ -128,15 +138,17 @@ async function main() {
     "AccessController",
     accessController.address,
     [],
+    "",
     contracts
   );
-  deploymentComplete("DiamondLoupeFacet", dlf.address, [], contracts);
-  deploymentComplete("DiamondCutFacet", dcf.address, [], contracts);
-  deploymentComplete("ERC165Facet", erc165f.address, [], contracts);
+  deploymentComplete("DiamondLoupeFacet", dlf.address, [], "", contracts);
+  deploymentComplete("DiamondCutFacet", dcf.address, [], "", contracts);
+  deploymentComplete("ERC165Facet", erc165f.address, [], "", contracts);
   deploymentComplete(
     "ProtocolDiamond",
     protocolDiamond.address,
     diamondArgs,
+    "",
     contracts
   );
 
@@ -152,6 +164,7 @@ async function main() {
     "ConfigHandlerFacet",
     configHandlerFacet.address,
     [],
+    "",
     contracts
   );
 
@@ -167,6 +180,7 @@ async function main() {
       deployedFacet.name,
       deployedFacet.contract.address,
       [],
+      "",
       contracts
     );
   }
@@ -194,18 +208,21 @@ async function main() {
     "BosonVoucher Logic",
     bosonVoucherImpl.address,
     [],
+    "",
     contracts
   );
   deploymentComplete(
     "BosonVoucher Beacon",
     bosonClientBeacon.address,
     [],
+    "",
     contracts
   );
   deploymentComplete(
     "BosonVoucher Proxy",
     bosonVoucherProxy.address,
     bosonVoucherProxyArgs,
+    "",
     contracts
   );
 
@@ -253,9 +270,23 @@ async function main() {
         "Foreign721",
         "Foreign1155"
       ]);
-    deploymentComplete("Foreign20", foreign20Token.address, [], contracts);
-    deploymentComplete("Foreign721", foreign721Token.address, [], contracts);
-    deploymentComplete("Foreign1155", foreign1155Token.address, [], contracts);
+    deploymentComplete("Foreign20", foreign20Token.address, [], "", contracts);
+    deploymentComplete(
+      "Foreign721",
+      foreign721Token.address,
+      [],
+      "",
+      contracts
+    );
+    deploymentComplete(
+      "Foreign1155",
+      foreign1155Token.address,
+      [],
+      "",
+      contracts
+    );
+
+    await deployAndMintMockNFTAuthTokens();
 
     // Create and activate default dispute resolver
     const accountHandler = await ethers.getContractAt(