+++ packages/core-sdk/src/accounts/handler.ts
@@ -14,7 +14,9 @@ import {
   encodeAddSellersToAllowList,
   encodeRemoveFeesFromDisputeResolver,
   encodeRemoveSellersFromAllowList,
-  encodeUpdateDisputeResolver
+  encodeUpdateDisputeResolver,
+  encodeOptInToSellerUpdate,
+  encodeOptInToDisputeResolverUpdate
 } from "./interface";
 import { getDisputeResolverById } from "./subgraph";
 import {
@@ -22,7 +24,9 @@ import {
   UpdateSellerArgs,
   CreateDisputeResolverArgs,
   DisputeResolutionFee,
-  DisputeResolverUpdates
+  DisputeResolverUpdates,
+  OptInToSellerUpdateArgs,
+  OptInToDisputeResolverUpdateArgs
 } from "./types";
 
 export async function createSeller(args: {
@@ -44,6 +48,17 @@ export async function updateSeller(args: {
   return args.web3Lib.sendTransaction({
     to: args.contractAddress,
     data: encodeUpdateSeller(args.sellerUpdates)
+  });
+}
+
+export async function optInToSellerUpdate(args: {
+  sellerUpdates: OptInToSellerUpdateArgs;
+  contractAddress: string;
+  web3Lib: Web3LibAdapter;
+}): Promise<TransactionResponse> {
+  return args.web3Lib.sendTransaction({
+    to: args.contractAddress,
+    data: encodeOptInToSellerUpdate(args.sellerUpdates)
   });
 }
 
@@ -146,6 +161,17 @@ export async function updateDisputeResolver(args: {
         ? utils.timestamp.msToSec(args.updates.escalationResponsePeriodInMS)
         : disputeResolver.escalationResponsePeriod
     })
+  });
+}
+
+export async function optInToDisputeResolverUpdate(args: {
+  disputeResolverUpdates: OptInToDisputeResolverUpdateArgs;
+  contractAddress: string;
+  web3Lib: Web3LibAdapter;
+}): Promise<TransactionResponse> {
+  return args.web3Lib.sendTransaction({
+    to: args.contractAddress,
+    data: encodeOptInToDisputeResolverUpdate(args.disputeResolverUpdates)
   });
 }