+++ packages/core-sdk/src/erc1155/interface.ts
@@ -1,4 +1,4 @@
 import { abis } from "@bosonprotocol/common";
 import { Interface } from "@ethersproject/abi";
 
-export const erc1155Iface = new Interface(abis.ERC1155);
+export const erc1155Iface = new Interface(abis.ERC1155ABI);