+++ packages/react-kit/src/hooks/useCtaClickHandler.ts
@@ -1,6 +1,6 @@
 import { TransactionResponse } from "@bosonprotocol/common";
-import { providers, errors } from "ethers";
 import { CoreSDK, metaTx } from "@bosonprotocol/core-sdk";
+import { errors, providers } from "ethers";
 import { useState } from "react";
 
 import { CtaButtonProps } from "../components/cta/common/types";
@@ -8,7 +8,9 @@ import { CtaButtonProps } from "../components/cta/common/types";
 type WriteContractFn = () => Promise<TransactionResponse>;
 type SignMetaTxFn = () => Promise<metaTx.handler.SignedMetaTx>;
 type MetaTxCondition = boolean;
+export type ActionName = "approveExchangeToken" | "depositFunds" | "commit";
 export type Action = {
+  name?: ActionName;
   signMetaTxFn?: SignMetaTxFn;
   writeContractFn: WriteContractFn;
   additionalMetaTxCondition?: MetaTxCondition;
@@ -57,6 +59,7 @@ export function useCtaClickHandler<T>({
 
       for (const action of actions) {
         const {
+          name,
           signMetaTxFn,
           writeContractFn,
           additionalMetaTxCondition = true,
@@ -72,7 +75,7 @@ export function useCtaClickHandler<T>({
           Boolean(coreSdk.isMetaTxConfigSet && signerAddress && signMetaTxFn) &&
           additionalMetaTxCondition;
 
-        onPendingSignature?.();
+        onPendingSignature?.(name);
 
         if (isMetaTx && signMetaTxFn) {
           const nonce = Date.now();
@@ -101,7 +104,7 @@ export function useCtaClickHandler<T>({
 
         if (txResponse) {
           try {
-            onPendingTransaction?.(txResponse.hash, isMetaTx);
+            onPendingTransaction?.(txResponse.hash, isMetaTx, name);
             receipt = await txResponse.wait(waitBlocks);
           } catch (error: any) {
             // Handle transaction that was replaced, canceled or repriced.
@@ -125,7 +128,8 @@ export function useCtaClickHandler<T>({
                   txResponse.hash,
                   error.replacement,
                   error.receipt,
-                  isMetaTx
+                  isMetaTx,
+                  name
                 );
                 receipt = error.receipt;
               }