+++ scripts/update-seller.ts
@@ -1,3 +1,4 @@
+import { AddressZero } from "@ethersproject/constants";
 import fs from "fs";
 import { EnvironmentType } from "@bosonprotocol/common/src/types/configs";
 import { providers, Contract, Wallet } from "ethers";
@@ -5,7 +6,7 @@ import { program } from "commander";
 import { getDefaultConfig, UpdateSellerArgs } from "@bosonprotocol/common/src";
 import { CoreSDK } from "../packages/core-sdk/src";
 import { EthersAdapter } from "../packages/ethers-sdk/src";
-import { abis } from "@bosonprotocol/common";
+import { abis, TransactionResponse } from "@bosonprotocol/common";
 import { extractSellerData } from "./utils/account";
 
 program
@@ -23,6 +24,10 @@ program
   .option("--operator <OPERATOR>", "New operator address")
   .option("--authTokenId <AUTH_TOKEN_ID>", "New Auth Token Id")
   .option("--authTokenType <AUTH_TOKEN_TYPE>", "New Auth Token Type")
+  .option(
+    "--privateKeys <PRIVATE_KEYS>",
+    "Comma-separated list of private keys used for opting in the update"
+  )
   .parse(process.argv);
 
 async function main() {
@@ -43,6 +48,41 @@ async function main() {
   const sellerId = opts.id || sellerDataJson.id;
   if (sellerId === undefined) {
     throw `Seller ID is required`;
+  }
+
+  const optInSigners = new Map<
+    string,
+    {
+      sdk: CoreSDK;
+      fields: {
+        admin: boolean;
+        operator: boolean;
+        clerk: boolean;
+        authToken: boolean;
+      };
+    }
+  >();
+  if (opts.privateKeys) {
+    const defaultFields = {
+      admin: false,
+      operator: false,
+      clerk: false,
+      authToken: false
+    };
+    (opts.privateKeys as string).split(",").forEach((privKey) => {
+      const wallet = new Wallet(privKey);
+      const sdk = CoreSDK.fromDefaultConfig({
+        web3Lib: new EthersAdapter(
+          new providers.JsonRpcProvider(defaultConfig.jsonRpcUrl),
+          wallet
+        ),
+        envName
+      });
+      optInSigners.set(wallet.address.toLowerCase(), {
+        sdk,
+        fields: { ...defaultFields } // clone the structure
+      });
+    });
   }
 
   const accountAbi = abis.IBosonAccountHandlerABI;
@@ -103,10 +143,56 @@ async function main() {
   });
 
   console.log(`Updating seller on env ${envName} on chain ${chainId}...`);
-  const txResponse1 = await coreSDK.updateSeller(sellerDataJson);
+  const txResponse1 = await coreSDK.updateSellerAndOptIn(sellerDataJson);
   console.log(`Tx hash: ${txResponse1.hash}`);
-  await txResponse1.wait();
-  console.log(`Seller with id ${sellerId} updated.`);
+  const txReceipt1 = await txResponse1.wait();
+  const pendingSellerUpdate = coreSDK.getPendingSellerUpdateFromLogs(
+    txReceipt1.logs
+  );
+  let updateComplete = true;
+  console.log(`Pending Seller Updates: ${JSON.stringify(pendingSellerUpdate)}`);
+  for (const key of ["operator", "clerk", "admin", "tokenType"]) {
+    if (pendingSellerUpdate[key] && pendingSellerUpdate[key] !== AddressZero) {
+      const address = pendingSellerUpdate[key].toLowerCase();
+      if (optInSigners.has(address)) {
+        (optInSigners.get(address) as any).fields[key] = true;
+      } else {
+        console.warn(
+          `No private key found to optIn for account ${address}. Update won't be complete`
+        );
+        updateComplete = false;
+      }
+    }
+  }
+  const txOptIns: TransactionResponse[] = [];
+  for (const [account, optInSigner] of optInSigners.entries()) {
+    if (
+      optInSigner.fields.admin ||
+      optInSigner.fields.operator ||
+      optInSigner.fields.clerk ||
+      optInSigner.fields.authToken
+    ) {
+      console.log(
+        `OptIn with account ${account} for fields ${JSON.stringify(
+          optInSigner.fields
+        )}`
+      );
+      txOptIns.push(
+        await optInSigner.sdk.optInToSellerUpdate({
+          id: sellerDataJson.id,
+          fieldsToUpdate: optInSigner.fields
+        })
+      );
+    }
+  }
+  await Promise.all(txOptIns.map((txOptIn) => txOptIn.wait()));
+  if (updateComplete) {
+    console.log(`Seller with id ${sellerId} updated.`);
+  } else {
+    console.warn(
+      `Seller with id ${sellerId} incomplete (you need to opt in for some accounts).`
+    );
+  }
 }
 
 main()