+++ e2e/meta-tx-gateway/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.4](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.2.3...meta-tx-gateway@1.2.4) (2022-12-01)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
+## [1.2.3](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.2.2...meta-tx-gateway@1.2.3) (2022-11-22)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
+## [1.2.2](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.2.1...meta-tx-gateway@1.2.2) (2022-11-17)
+
+**Note:** Version bump only for package meta-tx-gateway
+
+
+
+
+
 ## [1.2.1](https://github.com/bosonprotocol/core-components/compare/meta-tx-gateway@1.2.0...meta-tx-gateway@1.2.1) (2022-10-28)
 
 **Note:** Version bump only for package meta-tx-gateway