+++ packages/eth-connect-sdk/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.4](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/eth-connect-sdk@1.4.3...@bosonprotocol/eth-connect-sdk@1.4.4) (2022-12-01)
+
+**Note:** Version bump only for package @bosonprotocol/eth-connect-sdk
+
+
+
+
+
+## [1.4.3](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/eth-connect-sdk@1.4.2...@bosonprotocol/eth-connect-sdk@1.4.3) (2022-11-22)
+
+**Note:** Version bump only for package @bosonprotocol/eth-connect-sdk
+
+
+
+
+
+## [1.4.2](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/eth-connect-sdk@1.4.1...@bosonprotocol/eth-connect-sdk@1.4.2) (2022-11-17)
+
+**Note:** Version bump only for package @bosonprotocol/eth-connect-sdk
+
+
+
+
+
 ## [1.4.1](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/eth-connect-sdk@1.4.0...@bosonprotocol/eth-connect-sdk@1.4.1) (2022-10-28)
 
 **Note:** Version bump only for package @bosonprotocol/eth-connect-sdk