+++ packages/ipfs-storage/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.10.3](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ipfs-storage@1.10.2...@bosonprotocol/ipfs-storage@1.10.3) (2022-12-01)
+
+**Note:** Version bump only for package @bosonprotocol/ipfs-storage
+
+
+
+
+
+## [1.10.2](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ipfs-storage@1.10.1...@bosonprotocol/ipfs-storage@1.10.2) (2022-11-17)
+
+**Note:** Version bump only for package @bosonprotocol/ipfs-storage
+
+
+
+
+
 ## [1.10.1](https://github.com/bosonprotocol/core-components/compare/@bosonprotocol/ipfs-storage@1.10.0...@bosonprotocol/ipfs-storage@1.10.1) (2022-10-28)
 
 **Note:** Version bump only for package @bosonprotocol/ipfs-storage